backend/requirements.txt
crewai>=1.0.0
crewai-tools>=0.1.6
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
pydantic==2.5.3
openai>=1.12.0
httpx>=0.27.0
